Refactor Home page into sections
- Begin transition to TitleCase for component files
