+++ app/client/src/app/views/home/faq/Faq.scss
@@ -1,0 +1,14 @@
+@import "globals/hack-styles.scss";
+
+.faq-block {
+  width: 60%;
+  margin-bottom: 10rem;
+
+  @media screen and (max-width: $break-large) {
+    width: 80%;
+  }
+
+  h2 {
+    margin-bottom: 2rem;
+  }
+}
