+++ app/client/src/app/views/home/sponsor-info/CurrentPartners.js
