+++ app/client/src/app/views/home/sponsor-info/CurrentSponsors.js
