+++ app/client/src/app/views/index.js
@@ -2,7 +2,7 @@ export { default as Application } from "./application/application";
 export { default as Apply } from "./apply/apply";
 export { default as Confirmation } from "./confirmation/confirmation";
 export { default as Dashboard } from "./dashboard/dashboard";
-export { default as Home } from "./home/home";
+export { default as Home } from "./home/Home";
 export { default as Login } from "./login/login";
 export { default as LiveExpo } from "./live-expo/liveExpo";
 export { default as Schedule } from "./schedule/schedule";
